Add WS events handlers for delete message Add API description

+++ src/talk/talks.controller.ts
@@ -1,11 +1,21 @@
-import { Body, Controller, Delete, Get, Param, Post, Query, Req, UseGuards } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Query,
+  Req,
+  UseGuards,
+} from '@nestjs/common';
 import { TalkListQuery, TalkListQueryExtended } from '../models/query';
-import { from, Observable } from 'rxjs';
+import { from, Observable, ObservedValueOf } from 'rxjs';
 import { Talk, TalkType } from './entities/talk.entity';
 import { TalksService } from './services/talks.service';
 import { JwtAuthGuard } from '../auth/jwt-auth.guard';
-import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
-import { CreateMessageBody, RemoveMessageData, TalkData, UserTalkDeleteData } from '../models/talk';
+import { ApiBearerAuth, ApiOperation, ApiTags } from '@nestjs/swagger';
+import { CreateMessageBody, RemoveMessageData, TalkData } from '../models/talk';
 import { UsersService } from '../users-shared/users.service';
 import { UserStatus, UserTalk } from './entities/user-talk.entity';
 import { UserTalksService } from './services/user-talks.service';
@@ -25,11 +35,13 @@ export class TalksController {
   ) {}
 
   @Get('')
+  @ApiOperation({ summary: 'Find all talks' })
   findAll(@Query() query: TalkListQueryExtended): Observable<Talk[]> {
     return from(this.talksService.findWithParams(query));
   }
 
   @Get('/public')
+  @ApiOperation({ summary: 'Find all public talks' })
   findAllPublic(@Query() query: TalkListQuery): Observable<Talk[]> {
     return from(
       this.talksService.findWithParams({ ...query, type: TalkType.public }),
@@ -37,22 +49,40 @@ export class TalksController {
   }
 
   @Get(':talkId')
+  @ApiOperation({ summary: 'Get talk by ID' })
   getTalk(@Param('talkId') talkId: string): Observable<Talk> {
     return from(this.talksService.findOne(talkId));
   }
 
   @Post('')
-  async create(@Body() talk: TalkData, @Req() request) {
+  @ApiOperation({ summary: 'Create talk' })
+  async createTalk(@Body() talk: TalkData, @Req() request) {
     const userId = request.user.id;
     return this.userTalksService.createWithUser(talk, userId);
   }
 
   @Delete(':talkId')
-  async removeTalk(@Param('talkId') talkId: string) {
-    return this.talksService.delete(talkId);
+  @ApiOperation({
+    summary: `Remove talk. Talk will be removed if this is private talk or deleted by admin.
+      Otherwise user will be only removed from talk`,
+  })
+  async removeTalk(@Param('talkId') talkId: string, @Req() request) {
+    const talk = await this.talksService.findOne(talkId);
+    const userId = request.user.id;
+    if (talk.type === TalkType.private) {
+      return this.talksService.delete(talkId);
+    }
+    const userInTalk = await this.talksService.getUserInTalk(talkId, userId);
+    if (userInTalk && userInTalk.status === UserStatus.admin) {
+      return this.talksService.delete(talkId);
+    }
+    return this.userTalksService.removeUserFromTalk(talkId, userId);
   }
 
   @Post('conversation/:opponentId')
+  @ApiOperation({
+    summary: `Create conversation with user`,
+  })
   async createConversation(
     @Param('opponentId') opponentId: string,
     @Req() request,
@@ -61,7 +91,22 @@ export class TalksController {
     return this.userTalksService.createConversation(userId, opponentId);
   }
 
+  @Get('conversation/:opponentId')
+  @ApiOperation({
+    summary: `Get conversation with user`,
+  })
+  getConversation(
+    @Param('opponentId') opponentId: string,
+    @Req() request,
+  ): Observable<ObservedValueOf<Promise<UserTalk | null>>> {
+    const userId = request.user.id;
+    return from(this.talksService.getUserConversation(userId, opponentId));
+  }
+
   @Post(':talkId/user')
+  @ApiOperation({
+    summary: `Add user to talk`,
+  })
   async addUserToTalk(@Param('talkId') talkId: string, @Req() request) {
     const userId = request.user.id;
     await this.userTalksService.addUserToTalk(talkId, {
@@ -72,31 +117,39 @@ export class TalksController {
   }
 
   @Delete(':talkId/user/:userId')
+  @ApiOperation({
+    summary: `Remove user from talk`,
+  })
   async removeUserFromTalk(
     @Param('talkId') talkId: string,
-    @Body() userTalk: UserTalkDeleteData,
+    @Param('userId') userId: string,
   ) {
-    // const user = await this.userService.findOne(userTalk.userid);
-    // const talk = await this.talksService.findOne(userTalk.talkid);
-    // if (!user) {
-    //   throw new HttpException(`User not found`, HttpStatus.NOT_FOUND);
-    // }
-    // if (!talk) {
-    //   throw new HttpException(`User task not found`, HttpStatus.NOT_FOUND);
-    // }
-    // return this.talksService.addUserToTalk({
-    //   user,
-    //   talk,
-    //   status: userTalk.status,
-    // });
+    return this.userTalksService.removeUserFromTalk(talkId, userId);
   }
 
   @Get('user/talks')
+  @ApiOperation({
+    summary: `Get current user talks`,
+  })
   getUserTalks(@Req() request): Observable<UserTalk[]> {
     return from(this.talksService.getUserTalks(request.user.id));
   }
 
+  @Get(':talkId/user/:userId')
+  @ApiOperation({
+    summary: `Check if user in talk and get user status in talk`,
+  })
+  getTalkUser(
+    @Param('talkId') talkId: string,
+    @Param('userId') userId: string,
+  ): Observable<ObservedValueOf<Promise<UserTalk | null>>> {
+    return from(this.talksService.getUserInTalk(talkId, userId));
+  }
+
   @Get(':talkId/messages')
+  @ApiOperation({
+    summary: `Get talk messages`,
+  })
   getTalkMessages(
     @Param('talkId') talkId: string,
     @Req() request,
@@ -105,6 +158,9 @@ export class TalksController {
   }
 
   @Post(':talkId/messages')
+  @ApiOperation({
+    summary: `Add message to talk`,
+  })
   createTalkMessage(
     @Param('talkId') talkId: string,
     @Req() request,
@@ -120,6 +176,9 @@ export class TalksController {
   }
 
   @Delete(':talkId/messages')
+  @ApiOperation({
+    summary: `Remove message from talk`,
+  })
   async removeMessageFromTalk(
     @Param('talkId') talkId: string,
     @Req() request,
